refactor (post): move form content to separate component
